Added comments and some conditional logic canvas.js : 	refactored this file and seperated sprite creation and level generation into their own files. Added a function that is currently in development for collision detection (this function has two flavours, it checks if a collision is a concern with respect to a projectile and a sprite and then checks for collision if the the former is flagged as true.
levels.js
	added additional commentation to explain the use of this class
added conditonal logic to LevelParser.buildLevel so that only the first 10
characters of each comma seperated string passed in will be parsed
 ** this is temporary, but does prevent myself from unknowingly generating **
 ** more than 10 bricks on a row which would result in bricks not showing  **
 ** on the screen							   **

collisionDetection.js
	a crack at a collision detection, This program is in development.
the general idea is to narrow down the possability of a collision in two
phases like other collision detection schemes. The first phase will see
if a collision is possible dependant on the current projectiles trajectory
and the sprite of concerns position.