Continue work on collision detection function (see notes below)
Although there are many solutions available for collision detection (especially for a game of this nature) I have decided to take my own approach of solving this problem.
The reason for this is because it is a challenge that allows me to better understand physics and trig as well as better practice making good logic with javascript.